Updated Node, Server, System
Created Node and Server files, which separated the previous files. The
Node now has an add_parent and add_child method, which are fairly
trivial.

I also created a System.py file which can be used for the entire
environment. I created a method that has an add_connection that takes a
parent and child and called the respective add_parent and add_child
methods. Also, I created a decodable method that takes in a node, and
checks if the node has enough data to be decoded.